implement all backend functionality

+++ TNR_Backend/routes/comment-routes.js
@@ -1,4 +1,3 @@
-const jwtDecode = require('jwt-decode')
 const jwtService = require('../services/jwt-service')
 
 class CommentRoutes {
@@ -11,13 +10,19 @@ class CommentRoutes {
   addComment (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const comment = req.body.comment // comment
-      const post = req.body.post // post it belongs to
-      const user = req.body.user // which user added it
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const comment = req.body.comment // comment
+          const post = req.body.post // post it belongs to
+          const user = req.body.user // which user added it
 
-      return this.neo4jService.createComment(comment, user, post)
-        .then(resp => res.status(201).send({ status: 'Comment added' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4jService.createComment(comment, user, post)
+            .then(resp => res.status(201).send({ status: 'Comment added' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -26,11 +31,17 @@ class CommentRoutes {
   deleteComment (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const objToDelete = req.body // obj must have id
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const objToDelete = req.body // obj must have id
 
-      return this.neo4j.deleteNode('Comment', objToDelete)
-        .then(result => res.status(200).send({ status: 'Deleted' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4j.deleteNode('Comment', objToDelete)
+            .then(result => res.status(200).send({ status: 'Deleted' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -39,14 +50,20 @@ class CommentRoutes {
   updateComment (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const objectToUpdate = req.body // must have id
-      const idToFind = {
-        id: objectToUpdate.id
-      }
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const objectToUpdate = req.body // must have id
+          const idToFind = {
+            id: objectToUpdate.id
+          }
 
-      return this.neo4j.updateNode('Comment', idToFind, objectToUpdate)
-        .then(result => res.status(200).send({ status: 'Updated successfully.' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4j.updateNode('Comment', idToFind, objectToUpdate)
+            .then(result => res.status(200).send({ status: 'Updated successfully.' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -55,31 +72,18 @@ class CommentRoutes {
   getPostComments (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const token = header.slice(7)
-      const legit = jwtService.verifyToken(token)
-      // if (legit) {
-      return this.neo4j.selectRelationshipByNode2('Comment', 'Post', 'ON_POST', { id: req.body.post }) // get all node1(posts) by relationship
-        .then(result => res.status(200).send(result.records))
-        .catch(e => res.status(400).send(e))
-      // } else {
-      //   res.status(403).send({ error: 'Authorization Error. Access Denied.' })
-      // }
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          return this.neo4j.selectRelationshipByNode2('Comment', 'Post', 'ON_POST', { id: req.body.post }) // get all node1(posts) by relationship
+            .then(result => res.status(200).send(result.records))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
   }
-
-  // getAllUserCommentedPosts (req, res, next) {
-  //   const header = req.headers.authorization
-  //   if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-  //     const token = jwtDecode(header.slice(7))
-
-  //     return this.neo4j.selectRelationship3ByNode1('User', 'Comment', 'Post', token.userId, 'COMMENTED', 'ON_POST')
-  //       .then(result => res.status(200).send(result.records))
-  //       .catch(e => res.status(400).send(e))
-  //   } else {
-  //     res.status(403).send({ error: 'Authorization Error. Access Denied.' })
-  //   }
-  // }
 }
 module.exports.CommentRoutes = CommentRoutes
