+++ TNR_Backend/routes/community-routes.js
@@ -1,5 +1,4 @@
 const jwtService = require('../services/jwt-service')
-const jwtDecode = require('jwt-decode')
 
 class CommunityRoutes {
   constructor (neo4j, neo4jService, redis) {
@@ -11,12 +10,18 @@ class CommunityRoutes {
   addCommunity (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const community = req.body.community
-      const user = req.body.user // admin, who added it
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const community = req.body.community
+          const user = req.body.user // admin, who added it
 
-      return this.neo4jService.createCommunity(community, user)
-        .then(resp => res.status(201).send({ status: 'Community added' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4jService.createCommunity(community, user)
+            .then(resp => res.status(201).send({ status: 'Community added' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -25,11 +30,17 @@ class CommunityRoutes {
   deleteCommunity (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const objToDelete = req.body // obj must have id
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const objToDelete = req.body // obj must have id
 
-      return this.neo4j.deleteNode('Community', objToDelete)
-        .then(result => res.status(200).send({ status: 'Deleted' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4j.deleteNode('Community', objToDelete)
+            .then(result => res.status(200).send({ status: 'Deleted' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -38,14 +49,20 @@ class CommunityRoutes {
   updateCommunity (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const objectToUpdate = req.body // must have id
-      const idToFind = {
-        id: objectToUpdate.id
-      }
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const objectToUpdate = req.body // must have id
+          const idToFind = {
+            id: objectToUpdate.id
+          }
 
-      return this.neo4j.updateNode('Community', idToFind, objectToUpdate)
-        .then(result => res.status(200).send({ status: 'Updated successfully.' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4j.updateNode('Community', idToFind, objectToUpdate)
+            .then(result => res.status(200).send({ status: 'Updated successfully.' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -54,9 +71,15 @@ class CommunityRoutes {
   getAllCommunities (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      return this.neo4j.selectAllByLabel('Community')
-        .then(result => res.status(200).send(result.records))
-        .catch(e => res.status(400).send(e))
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          return this.neo4j.selectAllByLabel('Community')
+            .then(result => res.status(200).send(result.records))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
