+++ TNR_Backend/routes/post-routes.js
@@ -11,18 +11,17 @@ class PostRoutes {
   getAllCommunityPosts (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const token = header.slice(7)
-      // const legit = jwtService.verifyToken(token)
-      // if (legit) {
-      const community = {
-        title: req.params.community
-      }
-      return this.neo4j.selectRelationshipByNode2('Post', 'Community', 'SUBJECT', community) // get all node1(posts) by relationship
-        .then(result => res.status(200).send(result.records))
-        .catch(e => res.status(400).send(e))
-      // } else {
-      // res.status(403).send({ error: '3.' })
-      // }
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const community = req.params.community
+
+          return this.neo4jService.getAllCommunityPosts(community)
+            .then(result => res.status(200).send(result))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -31,13 +30,19 @@ class PostRoutes {
   addPost (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const post = req.body.post // post
-      const community = req.body.community // community it belongs to
-      const user = req.body.user // which user added it
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const post = req.body.post // post
+          const community = req.body.community // community it belongs to
+          const user = req.body.user // which user added it
 
-      return this.neo4jService.createPost(post, user, community)
-        .then(resp => res.status(201).send({ status: 'Post created' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4jService.createPost(post, user, community)
+            .then(resp => res.status(201).send({ status: 'Post created' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -46,11 +51,17 @@ class PostRoutes {
   deletePost (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const objToDelete = req.body // obj must have id
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const objToDelete = req.body // obj must have id
 
-      return this.neo4j.deleteNode('Post', objToDelete)
-        .then(result => res.status(200).send({ status: 'Deleted' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4j.deleteNode('Post', objToDelete)
+            .then(result => res.status(200).send({ status: 'Deleted' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -59,14 +70,20 @@ class PostRoutes {
   updatePost (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const objectToUpdate = req.body // must have id
-      const idToFind = {
-        id: objectToUpdate.id
-      }
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const objectToUpdate = req.body // must have id
+          const idToFind = {
+            id: objectToUpdate.id
+          }
 
-      return this.neo4j.updateNode('Post', idToFind, objectToUpdate)
-        .then(result => res.status(200).send({ status: 'Updated successfully.' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4j.updateNode('Post', idToFind, objectToUpdate)
+            .then(result => res.status(200).send({ status: 'Updated successfully.' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -75,15 +92,21 @@ class PostRoutes {
   upVote (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const decodedToken = jwtDecode(header.slice(7))
-      const userId = {
-        id: decodedToken.userId
-      }
-      const post = req.body.post
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const decodedToken = jwtDecode(header.slice(7))
+          const userId = {
+            id: decodedToken.userId
+          }
+          const post = req.body.post
 
-      return this.neo4jService.postUpVote(userId.id, userId, post)
-        .then(resp => res.status(200).send({ status: 'Post votes updated.' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4jService.postUpVote(userId.id, userId, post)
+            .then(resp => res.status(200).send({ status: 'Post votes updated.' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -92,26 +115,38 @@ class PostRoutes {
   downVote (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      const decodedToken = jwtDecode(header.slice(7))
-      const userId = {
-        id: decodedToken.userId
-      }
-      const post = req.body.post
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          const decodedToken = jwtDecode(header.slice(7))
+          const userId = {
+            id: decodedToken.userId
+          }
+          const post = req.body.post
 
-      return this.neo4jService.postDownVote(userId.id, userId, post)
-        .then(resp => res.status(200).send({ status: 'Post votes updated.' }))
-        .catch(e => res.status(400).send(e))
+          return this.neo4jService.postDownVote(userId.id, userId, post)
+            .then(resp => res.status(200).send({ status: 'Post votes updated.' }))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
   }
 
-  getMostLikedPost (req, res, next) {
+  getMostLikedPosts (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      return this.neo4j.getNode1WithMaxRelationship('Post', 'User', 'UPVOTED_BY')
-        .then(resp => res.status(200).send(resp))
-        .catch(e => res.status(400).send(e))
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          return this.neo4j.getNode1WithMaxRelationship('Post', 'User', 'UPVOTED_BY')
+            .then(resp => res.status(200).send(resp))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
@@ -120,9 +155,15 @@ class PostRoutes {
   getCommunity (req, res, next) {
     const header = req.headers.authorization
     if (header && (header.indexOf('Bearer') !== -1 || header.indexOf('bearer') !== -1) && header.indexOf('undefined') === -1) {
-      return this.neo4j.selectRelationshipByNode2('Community', 'Post', 'HAS_POST', req.body)
-        .then(resp => res.status(200).send(resp))
-        .catch(e => res.status(400).send(e))
+      jwtService.jwtVerify(header.slice(7), req.session.userId).subscribe(result => {
+        if (result) {
+          return this.neo4j.selectRelationshipByNode2('Community', 'Post', 'HAS_POST', req.body)
+            .then(resp => res.status(200).send(resp))
+            .catch(e => res.status(400).send(e))
+        } else {
+          res.status(403).send({ error: 'Authorization Error. Access Denied.' })
+        }
+      })
     } else {
       res.status(403).send({ error: 'Authorization Error. Access Denied.' })
     }
